S3900: Dummy rule implementation (#6917)
